requirements.txt
streamlit==1.44.1
langchain==0.1.14
google-generativeai==0.8.4
python-dotenv==1.1.0
deep-translator==1.11.4
chromadb==1.0.4
PyPDF2==3.0.1

